libs/core/pyproject.toml
[project]
name = "vizier-core"
version = "0.1.0"
description = "Autonomous multi-agent work system - Core library"
requires-python = ">=3.11,<3.13"
dependencies = [
    "litellm>=1.0",
    "pydantic>=2.0",
    "jinja2>=3.0",
    "watchdog>=6.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.0",
    "filelock>=3.0",
    "structlog>=24.0",
    "azure-identity>=1.15",
    "azure-keyvault-secrets>=4.8",
]

[dependency-groups]
dev = [
    "pyright>=1.1.390",
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "standard"
venvPath = "../.."
venv = ".venv"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["vizier"]
